third_party/WebKit/LayoutTests/fast/css/invalidation/targeted-class-style-invalidation.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>

<style>
  div { width: 100px }
  .outer .inner { width: 200px }
</style>

<div id="outer">
    <div id="inner" class="inner">
        <div id="innerChild">
        </div>
    </div>
    <div id="inner2">
    </div>
</div>

<script>
description("Test that adding and removing class names only updates the elements that are affected.");

function insertStyleSheet(css)
{
    var styleElement = document.createElement("style");
    styleElement.textContent = css;
    (document.head || document.documentElement).appendChild(styleElement);
}

var outer = document.getElementById('outer');
var inner = document.getElementById('inner');

var count;
if (internals.runtimeFlags.targetedStyleRecalcEnabled)
    count = 2;
else
    count = 4;

// Style recalc should happen on "inner" and "outer", but not "inner2".
outer.offsetTop;
outer.className = 'outer';
shouldBe("internals.updateStyleAndReturnAffectedElementCount()", '' + count);
shouldBe("getComputedStyle(inner).width", '"200px"');

if (internals.runtimeFlags.targetedStyleRecalcEnabled)
    count = 1;
else
    count = 2;

// Style recalc should happen on "inner", but not "innerChild".
inner.offsetTop;
inner.className = '';
shouldBe("internals.updateStyleAndReturnAffectedElementCount()", '' + count);
shouldBe("getComputedStyle(inner).width", '"100px"');

</script>
